requirements.txt
pymongo==3.12.3
Django==4.1.4
djongo==1.3.6
pytz==2022.7
Markdown==3.2.1
Pillow==9.0.1
dnspython==2.2.1
django-tinymce=3.5.0